Added functionality for semi-approved domains
This will allow us to continue to create new accounts on the fly for CAST and Catalyst email holders, but only allow SSO for other domains if the user already has an account. So we could for example approve the gmail.com domain allowing people with that domain to use SSO, but only if they already have an account.